+++ backend/tests/app/role_assignments/data_product/test_router.py
@@ -9,7 +9,7 @@ from tests.factories import (
 )
 from tests.factories.data_product_membership import DataProductMembershipFactory
 
-from app.core.authz.actions import AuthorizationAction
+from app.core.authz import Action
 from app.data_products.model import DataProduct
 from app.role_assignments.data_product.schema import RoleAssignment
 from app.role_assignments.enums import DecisionStatus
@@ -18,6 +18,7 @@ from app.users.schema import User
 
 ENDPOINT = "/api/role_assignments/data_product"
 ENDPOINT_DATA_PRODUCT = "/api/data_products"
+ENDPOINT_PENDING_ACTIONS = "/api/pending_actions"
 
 
 class TestDataProductRoleAssignmentsRouter:
@@ -221,6 +222,51 @@ class TestDataProductRoleAssignmentsRouter:
         assert response.status_code == 200
         data = response.json()
         assert len(data) == 0
+
+    def test_get_pending_actions_no_action(self, client):
+        user = UserFactory(external_id="sub")
+        data_product: DataProduct = DataProductMembershipFactory(
+            user=user,
+        ).data_product
+        role: Role = RoleFactory(scope=Scope.DATA_PRODUCT)
+        DataProductRoleAssignmentFactory(
+            data_product_id=data_product.id,
+            user_id=user.id,
+            role_id=role.id,
+            decision=DecisionStatus.APPROVED,
+        )
+        response = client.get(f"{ENDPOINT_PENDING_ACTIONS}")
+        assert response.json() == []
+
+    def test_request_data_product_role_assignment(self, client):
+        data_product: DataProduct = DataProductFactory()
+        user: User = UserFactory(external_id="sub")
+        role: Role = RoleFactory(
+            scope=Scope.DATA_PRODUCT,
+            permissions=[Action.DATA_PRODUCT__APPROVE_USER_REQUEST],
+        )
+        DataProductRoleAssignmentFactory(
+            data_product_id=data_product.id,
+            user_id=user.id,
+            role_id=role.id,
+            decision=DecisionStatus.APPROVED,
+        )
+        user_requester: User = UserFactory()
+        role: Role = RoleFactory(scope=Scope.DATA_PRODUCT)
+
+        response = client.post(
+            f"{ENDPOINT}",
+            json={
+                "data_product_id": str(data_product.id),
+                "user_id": str(user_requester.id),
+                "role_id": str(role.id),
+            },
+        )
+        assert response.status_code == 200
+
+        response = client.get(f"{ENDPOINT_PENDING_ACTIONS}")
+        assert response.status_code == 200
+        assert len(response.json()) == 1
 
     @staticmethod
     def delete_data_product(client, data_product_id):
